fix(auth): enforce authentication before purchase
