+++ package.json
@@ -11,7 +11,7 @@
   "dependencies": {
     "@iconify/icons-ion": "^1.2.10",
     "@iconify/react": "^5.0.1",
-    "@prisma/client": "^6.4.1",
+    "@prisma/client": "^6.5.0",
     "@stripe/stripe-js": "^6.0.0",
     "aos": "^2.3.4",
     "axios": "^1.7.2",
@@ -23,7 +23,8 @@
     "next-auth": "^4.24.11",
     "next-themes": "^0.3.0",
     "nextjs-toploader": "^3.7.15",
-    "prisma": "^6.4.1",
+    "prism": "^4.1.2",
+    "prisma": "^6.5.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hot-toast": "^2.4.1",
