+++ src/app/api/checkout-session/route.js
@@ -1,8 +1,47 @@
+// import Stripe from "stripe";
+
+// const stripe = new Stripe(process.env.STRIPE_SECRET_KEY);
+
+// export async function POST(req) {
+//   try {
+//     const { priceId } = await req.json();
+
+//     if (!priceId) {
+//       return new Response(JSON.stringify({ error: "Price ID is required" }), { status: 400 });
+//     }
+
+//     const session = await stripe.checkout.sessions.create({
+//       payment_method_types: ["card"],
+//       mode: "payment",
+//       line_items: [
+//         {
+//           price: priceId,
+//           quantity: 1,
+//         },
+//       ],
+//       success_url: `${process.env.NEXT_PUBLIC_BASE_URL}/success`,
+//       cancel_url: `${process.env.NEXT_PUBLIC_BASE_URL}/cancel`,
+//     });
+
+//     return new Response(JSON.stringify({ url: session.url }), { status: 200 });
+//   } catch (error) {
+//     return new Response(JSON.stringify({ error: error.message }), { status: 500 });
+//   }
+// }
+
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { getServerSession } from "next-auth";
 import Stripe from "stripe";
 
 const stripe = new Stripe(process.env.STRIPE_SECRET_KEY);
 
 export async function POST(req) {
+  // Get user session
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    return new Response(JSON.stringify({ error: "Unauthorized: Please sign in first" }), { status: 401 });
+  }
+
   try {
     const { priceId } = await req.json();
 
@@ -10,8 +49,12 @@ export async function POST(req) {
       return new Response(JSON.stringify({ error: "Price ID is required" }), { status: 400 });
     }
 
-    const session = await stripe.checkout.sessions.create({
+    // Retrieve user email from session
+    const userEmail = session.user.email;
+
+    const stripeSession = await stripe.checkout.sessions.create({
       payment_method_types: ["card"],
+      customer_email: userEmail, // Attach user email to Stripe session
       mode: "payment",
       line_items: [
         {
@@ -23,7 +66,7 @@ export async function POST(req) {
       cancel_url: `${process.env.NEXT_PUBLIC_BASE_URL}/cancel`,
     });
 
-    return new Response(JSON.stringify({ url: session.url }), { status: 200 });
+    return new Response(JSON.stringify({ url: stripeSession.url }), { status: 200 });
   } catch (error) {
     return new Response(JSON.stringify({ error: error.message }), { status: 500 });
   }
